requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9  
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0             
numpy>=1.26.0
pdfplumber>=0.11.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.2.0
openai>=1.30.0
streamlit>=1.35.0